integer price in db model

+++ backend/app/models/receiptline.py
@@ -1,5 +1,5 @@
 from sqlalchemy.sql.schema import Column, ForeignKey
-from sqlalchemy.sql.sqltypes import Integer, Numeric, Text
+from sqlalchemy.sql.sqltypes import Integer, Integer, Text
 
 from backend.app.db.base_class import Base
 
@@ -10,4 +10,4 @@ class Receiptline(Base):
     receipt_id = Column(Text, ForeignKey('receipt.id'), primary_key=True)
     linenumber = Column(Integer, primary_key=True)
     product_id = Column(Text, ForeignKey('product.id'))
-    amount = Column(Numeric)
+    amount = Column(Integer)
